Do not fail if resource table has not been found.
While most normal applications have such a resource table, several system
applications don't, e.g. the Smoke-Test app pre-installed on the emulator
(can be found at /data/app/SmokeTestApp.apk). If an application is missing
a resource table, it is a valid assumtion that it is not needed and
no resources will be referenced from e.g. the manifest file.

With this patch, ApkParser will not throw a ParserException by default if
a resource table has not been found. Instead, it will initialize the
resource table with an empty object, and the locale set with an empty set.
This fixes the bug with SmokeTest and other system apps.