Parse the debuggable information
Apps can be build in debug or in release mode. `./gradlew assembleDebug`
vs `./gradlew assembleRelease`. If the app is debuggable the Android OS
allows to attach the debugger and other development features. In release
mode these are restricted.

The AndroidManifest flag is documented here:

    https://developer.android.com/guide/topics/manifest/application-element#debug

Since this information is quite handy, add it to the parser.